backend/requirements.txt
alembic==1.14.0
annotated-types==0.7.0
anyio==4.7.0
asyncpg==0.30.0
authx==1.4.0
cffi==1.17.1
click==8.1.7
cryptography==44.0.0
ecdsa==0.19.0
fastapi==0.115.5
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Mako==1.3.8
MarkupSafe==3.0.2
psycopg==3.2.3
psycopg-binary==3.2.3
pyasn1==0.6.1
pycparser==2.22
pydantic==2.10.2
pydantic-settings==2.6.1
pydantic_core==2.27.1
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
pytz==2024.2
rsa==4.9
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
typing_extensions==4.12.2
uvicorn==0.32.1
